refactor action grid into ActionsPanel component
