+++ packages/web/src/components/HoverCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderSummary, renderCosts } from '../Game';
+import { renderSummary, renderCosts } from '../render';
 import { useGameEngine } from '../state/GameContext';
 
 type HoverCardData = ReturnType<typeof useGameEngine>['hoverCard'];
